refactored table alignment
