--- comet_pqc/tablecontrol.py
@@ -1,1510 +1,0 @@
-"""Table control widgets and dialogs."""
-
-import logging
-import math
-from typing import Optional
-
-from PyQt5 import QtCore, QtWidgets
-
-import comet
-import comet.ui as ui
-from PyQt5 import QtCore, QtGui, QtChart
-
-from .components import (
-    CalibrationWidget,
-    PositionLabel,
-    PositionWidget,
-    ToggleButton,
-)
-from .core.position import Position
-from .settings import TablePosition, settings
-from .utils import format_metric, caldone_valid, format_switch, handle_exception
-
-DEFAULT_STEP_UP_DELAY = 0.
-DEFAULT_STEP_UP_MULTIPLY = 2
-DEFAULT_LCR_UPDATE_INTERVAL = .100
-DEFAULT_MATRIX_CHANNELS = [
-    "3H01", "2B04", "1B03", "1B12", "2B06", "2B07", "2B08", "2B09",
-    "2B10", "2B11", "1H04", "1H05", "1H06", "1H07", "1H08", "1H09",
-    "1H10", "1H11", "2H12", "2H05", "1A01"
-]
-
-logger = logging.getLogger(__name__)
-
-
-def safe_z_position(z):
-    z_limit = settings.table_z_limit
-    if z > z_limit:
-        ui.show_warning(
-            title="Z Warning",
-            text=f"Limiting Z movement to {z_limit:.3f} mm to protect probe card."
-        )
-        z = z_limit
-    return z
-
-
-class LinearTransform:
-    """Linear transformation of n coordinates between two points."""
-
-    def calculate(self, a, b, n):
-        diff_x = (a[0] - b[0]) / n
-        diff_y = (a[1] - b[1]) / n
-        diff_z = (a[2] - b[2]) / n
-        return [(a[0] - diff_x * i, a[1] - diff_y * i, a[2] - diff_z * i) for i in range(n + 1)]
-
-
-class TableSampleItem(ui.TreeItem):
-
-    def __init__(self, sample_item):
-        super().__init__()
-        self.sample_item = sample_item
-        self.name = "/".join([item for item in (sample_item.name, sample_item.sample_type) if item])
-        self.position = sample_item.sample_position
-        for contact_item in sample_item.children:
-            self.append(TableContactItem(contact_item))
-
-    @property
-    def name(self):
-        return self[0].value
-
-    @name.setter
-    def name(self, value):
-        self[0].value = value
-
-    @property
-    def position(self):
-        return self[1].value
-
-    @position.setter
-    def position(self, value):
-        self[1].value = value
-
-    def update_contacts(self):
-        for contact_item in self.children:
-            contact_item.update_contact()
-            logger.info("Updated contact position: %s %s %s", self.name, contact_item.name, contact_item.position)
-
-    def calculate_positions(self):
-        tr = LinearTransform()
-        count = len(self.children)
-        if count > 2:
-            first = self.children[0].position
-            last = list(self.children)[-1].position
-            for i, position in enumerate(tr.calculate(first, last, count - 1)):
-                self.children[i].position = position
-
-
-class TableContactItem(ui.TreeItem):
-
-    def __init__(self, contact_item):
-        super().__init__()
-        for i in range(2, 5):
-            self[i].qt.setTextAlignment(i, QtCore.Qt.AlignTrailing | QtCore.Qt.AlignVCenter)
-        self.contact_item = contact_item
-        self.name = contact_item.name
-        self.position = contact_item.position
-
-    @property
-    def name(self):
-        return self[0].value
-
-    @name.setter
-    def name(self, value):
-        self[0].value = value
-
-    @property
-    def position(self):
-        return self.__position
-
-    @position.setter
-    def position(self, value):
-        x, y, z = value
-        self.__position = x, y, z
-        self[2].value = format(x, ".3f") if x is not None else None
-        self[3].value = format(y, ".3f") if y is not None else None
-        self[4].value = format(z, ".3f") if z is not None else None
-
-    @property
-    def has_position(self):
-        return any((not math.isnan(value) for value in self.__position))
-
-    def update_contact(self):
-        self.contact_item.position = self.position
-
-    def reset_position(self):
-        self.__position = float("nan"), float("nan"), float("nan")
-
-
-class TableContactsWidget(ui.Row):
-
-    def __init__(self, position_picked=None, absolute_move=None, **kwargs):
-        super().__init__(**kwargs)
-        self.position_picked = position_picked
-        self.absolute_move = absolute_move
-        self.contacts_tree = ui.Tree(
-            header=("Contact", "Pos", "X", "Y", "Z", None),
-            selected=self.on_contacts_selected
-        )
-        self.contacts_tree.fit()
-        self.pick_button = ui.Button(
-            text="Assign &Position",
-            tool_tip="Assign current table position to selected position item",
-            clicked=self.on_pick_position,
-            enabled=False
-        )
-        self.calculate_button= ui.Button(
-            text="&Calculate",
-            clicked=self.on_calculate,
-            enabled=False
-        )
-        self.move_button= ui.Button(
-            text="&Move",
-            tool_tip="Move to selected position",
-            clicked=self.on_move,
-            enabled=False
-        )
-        self.reset_button = ui.Button(
-            text="&Reset",
-            clicked=self.on_reset,
-            enabled=False
-        )
-        self.reset_all_button = ui.Button(
-            text="Reset &All",
-            clicked=self.on_reset_all
-        )
-        self.append(self.contacts_tree)
-        self.append(ui.Column(
-            self.pick_button,
-            self.move_button,
-            self.calculate_button,
-            ui.Spacer(),
-            self.reset_button,
-            self.reset_all_button
-        ))
-        self.stretch = 1, 0
-
-    def append_sample(self, sample_item):
-        self.contacts_tree.append(TableSampleItem(sample_item))
-
-    def on_contacts_selected(self, item):
-        self.update_button_states(item)
-
-    def update_button_states(self, item=None):
-        if item is None:
-            item = self.contacts_tree.current
-        is_contact = isinstance(item, TableContactItem)
-        self.pick_button.enabled = is_contact
-        self.move_button.enabled = item.has_position if is_contact else False
-        self.calculate_button.enabled = not is_contact
-        self.reset_button.enabled = is_contact
-
-    def on_pick_position(self):
-        item = self.contacts_tree.current
-        if isinstance(item, TableContactItem):
-            def callback(x, y, z):
-                item.position = x, y, z
-                self.contacts_tree.fit()
-            self.emit(self.position_picked, callback)
-
-    def on_reset(self):
-        item = self.contacts_tree.current
-        if isinstance(item, TableContactItem):
-            item.position = float("nan"), float("nan"), float("nan")
-            self.contacts_tree.fit()
-
-    def on_reset_all(self):
-        if ui.show_question("Do you want to reset all contact positions?"):
-            for sample_item in self.contacts_tree:
-                for contact_item in sample_item.children:
-                    contact_item.position = float("nan"), float("nan"), float("nan")
-            self.contacts_tree.fit()
-
-    def on_move(self):
-        current_item = self.contacts_tree.current
-        if isinstance(current_item, TableContactItem):
-            if current_item.has_position:
-                if ui.show_question(f"Do you want to move table to contact {current_item.name}?"):
-                    x, y, z = current_item.position
-                    self.emit(self.absolute_move, Position(x, y, z))
-
-    def on_calculate(self):
-        current_item = self.contacts_tree.current
-        if isinstance(current_item, TableSampleItem):
-            current_item.calculate_positions()
-
-    def loadSamples(self, sample_items) -> None:
-        self.contacts_tree.clear()
-        for sample_item in sample_items:
-            self.append_sample(sample_item)
-        self.contacts_tree.fit()
-
-    def updateSamples(self) -> None:
-        for sample_item in self.contacts_tree:
-            sample_item.update_contacts()
-
-    def setLocked(self, locked: bool) -> None:
-        if locked:
-            self.pick_button.qt.setEnabled(False)
-            self.move_button.qt.setEnabled(False)
-            self.calculate_button.qt.setEnabled(False)
-            self.reset_button.qt.setEnabled(False)
-        else:
-            self.update_button_states()
-        self.reset_all_button.qt.setEnabled(not locked)
-
-
-class TablePositionItem(ui.TreeItem):
-
-    def __init__(self, name, x, y, z, comment=None):
-        super().__init__()
-        for i in range(1, 4):
-            self[i].qt.setTextAlignment(i, QtCore.Qt.AlignTrailing | QtCore.Qt.AlignVCenter)
-        self.name = name
-        self.position = x, y, z
-        self.comment = comment or ""
-
-    @property
-    def name(self):
-        return self[0].value
-
-    @name.setter
-    def name(self, value):
-        self[0].value = value
-
-    @property
-    def position(self):
-        return self.__position
-
-    @position.setter
-    def position(self, value):
-        x, y, z = value
-        self.__position = x, y, z
-        self[1].value = format(x, ".3f")
-        self[2].value = format(y, ".3f")
-        self[3].value = format(z, ".3f")
-
-    @property
-    def comment(self):
-        return self[4].value
-
-    @comment.setter
-    def comment(self, value):
-        self[4].value = value
-
-
-class PositionDialog(QtWidgets.QDialog):
-
-    positionPicked = QtCore.pyqtSignal()
-
-    def __init__(self, parent: Optional[QtWidgets.QWidget] = None) -> None:
-        super().__init__(parent)
-
-        self.nameLabel: QtWidgets.QLabel = QtWidgets.QLabel(self)
-        self.nameLabel.setText("Name")
-        self.nameLabel.setToolTip("Position name")
-
-        self.nameLineEdit: QtWidgets.QLineEdit = QtWidgets.QLineEdit(self)
-        self.nameLineEdit.setText("Unnamed")
-
-        self.xLabel: QtWidgets.QLabel = QtWidgets.QLabel(self)
-        self.xLabel.setText("X")
-        self.xLabel.setToolTip("Position X coordinate")
-
-        self.xSpinBox: QtWidgets.QDoubleSpinBox = QtWidgets.QDoubleSpinBox(self)
-        self.xSpinBox.setRange(0, 1000)
-        self.xSpinBox.setDecimals(3)
-        self.xSpinBox.setSuffix(" mm")
-
-        self.yLabel: QtWidgets.QLabel = QtWidgets.QLabel(self)
-        self.yLabel.setText("X")
-        self.yLabel.setToolTip("Position Y coordinate")
-
-        self.ySpinBox: QtWidgets.QDoubleSpinBox = QtWidgets.QDoubleSpinBox(self)
-        self.ySpinBox.setRange(0, 1000)
-        self.ySpinBox.setDecimals(3)
-        self.ySpinBox.setSuffix(" mm")
-
-        self.zLabel: QtWidgets.QLabel = QtWidgets.QLabel(self)
-        self.zLabel.setText("Z")
-        self.zLabel.setToolTip("Position Z coordinate")
-
-        self.zSpinBox: QtWidgets.QDoubleSpinBox = QtWidgets.QDoubleSpinBox(self)
-        self.zSpinBox.setRange(0, 1000)
-        self.zSpinBox.setDecimals(3)
-        self.zSpinBox.setSuffix(" mm")
-
-        self.commentLabel: QtWidgets.QLabel = QtWidgets.QLabel(self)
-        self.commentLabel.setText("Comment")
-        self.commentLabel.setToolTip("Optional position comment")
-
-        self.commentLineEdit: QtWidgets.QLineEdit = QtWidgets.QLineEdit(self)
-
-        self.assignButton: QtWidgets.QPushButton = QtWidgets.QPushButton(self)
-        self.assignButton.setText("Assign Position")
-        self.assignButton.setToolTip("Assign current table position.")
-        self.assignButton.clicked.connect(self.positionPicked.emit)
-
-        self.buttonBox: QtWidgets.QDialogButtonBox = QtWidgets.QDialogButtonBox(self)
-        self.buttonBox.addButton(QtWidgets.QDialogButtonBox.Ok)
-        self.buttonBox.addButton(QtWidgets.QDialogButtonBox.Cancel)
-        self.buttonBox.accepted.connect(self.accept)
-        self.buttonBox.rejected.connect(self.reject)
-
-        gridLayout = QtWidgets.QGridLayout()
-        gridLayout.addWidget(self.xLabel, 0, 0)
-        gridLayout.addWidget(self.xSpinBox, 1, 0)
-        gridLayout.addWidget(self.yLabel, 0, 1)
-        gridLayout.addWidget(self.ySpinBox, 1, 1)
-        gridLayout.addWidget(self.zLabel, 0, 2)
-        gridLayout.addWidget(self.zSpinBox, 1, 2)
-        gridLayout.addWidget(self.assignButton, 1, 3)
-        gridLayout.setColumnStretch(3, 1)
-
-        layout = QtWidgets.QVBoxLayout(self)
-        layout.addWidget(self.nameLabel)
-        layout.addWidget(self.nameLineEdit)
-        layout.addLayout(gridLayout)
-        layout.addWidget(self.commentLabel)
-        layout.addWidget(self.commentLineEdit)
-        layout.addStretch()
-        layout.addWidget(self.buttonBox)
-
-    def name(self) -> None:
-        return self.nameLineEdit.text()
-
-    def setName(self, name: str) -> None:
-        self.nameLineEdit.setText(name)
-
-    def position(self) -> tuple:
-        x = self.xSpinBox.value()
-        y = self.ySpinBox.value()
-        z = self.zSpinBox.value()
-        return x, y, z
-
-    def setPosition(self, x: float, y: float, z: float) -> None:
-        self.xSpinBox.setValue(x)
-        self.ySpinBox.setValue(y)
-        self.zSpinBox.setValue(z)
-
-    def comment(self) -> str:
-        return self.commentLineEdit.text()
-
-    def setComment(self, comment: str) -> None:
-        self.commentLineEdit.setText(comment)
-
-
-class TablePositionsWidget(ui.Row):
-
-    def __init__(self, position_picked=None, absolute_move=None, **kwargs):
-        super().__init__(**kwargs)
-        self.position_picked = position_picked
-        self.absolute_move = absolute_move
-        self.positions_tree = ui.Tree(
-            header=("Name", "X", "Y", "Z", "Comment"),
-            root_is_decorated=False,
-            selected=self.on_position_selected,
-            double_clicked=self.on_position_double_clicked
-        )
-        self.add_button = ui.Button(
-            text="&Add",
-            tool_tip="Add new position item",
-            clicked=self.on_add_position
-        )
-        self.edit_button = ui.Button(
-            text="&Edit",
-            tool_tip="Edit selected position item",
-            clicked=self.on_edit_position,
-            enabled=False
-        )
-        self.remove_button = ui.Button(
-            text="&Remove",
-            tool_tip="Remove selected position item",
-            clicked=self.on_remove_position,
-            enabled=False
-        )
-        self.move_button= ui.Button(
-            text="&Move",
-            tool_tip="Move to selected position",
-            clicked=self.on_move,
-            enabled=False
-        )
-        self.import_button = ui.Button(
-            text="Import...",
-            tool_tip="Import from CSV",
-            clicked=self.on_import,
-        )
-        self.export_button = ui.Button(
-            text="Export...",
-            tool_tip="Export to CSV",
-            clicked=self.on_export,
-        )
-        self.append(self.positions_tree)
-        self.append(ui.Column(
-            self.move_button,
-            ui.Spacer(),
-            self.add_button,
-            self.edit_button,
-            self.remove_button,
-            self.import_button,
-            self.export_button,
-
-        ))
-        self.stretch = 1, 0
-
-    def readSettings(self) -> None:
-        self.positions_tree.clear()
-        for position in settings.table_positions:
-            self.positions_tree.append(TablePositionItem(
-                name=position.name,
-                x=position.x,
-                y=position.y,
-                z=position.z,
-                comment=position.comment,
-            ))
-        self.positions_tree.fit()
-
-    def writeSettings(self) -> None:
-        positions = []
-        for position in self.positions_tree:
-            x, y, z = position.position
-            positions.append(TablePosition(
-                name=position.name,
-                x=x,
-                y=y,
-                z=z,
-                comment=position.comment
-            ))
-        settings.table_positions = positions
-
-    def setLocked(self, locked: bool) -> None:
-        if locked:
-            self.add_button.enabled = False
-            self.edit_button.enabled = False
-            self.remove_button.enabled = False
-            self.move_button.enabled = False
-            # Remove event
-            self.positions_tree.double_clicked = None
-        else:
-            enabled = self.positions_tree.current is not None
-            self.add_button.enabled = True
-            self.edit_button.enabled = enabled
-            self.remove_button.enabled = enabled
-            self.move_button.enabled = enabled
-            # Restore event
-            self.positions_tree.double_clicked = self.on_position_double_clicked
-
-    def on_import(self) -> None:
-        filename, ok = QtWidgets.QFileDialog.getOpenFileName(self.qt, "Import CSV", "", "CSV (*.csv);;All files (*)")
-        if ok:
-            ...
-
-
-    def on_export(self) -> None:
-        filename, ok = QtWidgets.QFileDialog.getSaveFileName(self.qt, "Export CSV", "", "CSV (*.csv);;All files (*)")
-        if ok:
-            ...
-
-    def on_position_selected(self, item):
-        enabled = item is not None
-        self.edit_button.enabled = True
-        self.remove_button.enabled = True
-        self.move_button.enabled = True
-
-    def on_position_double_clicked(self, *args):
-        self.on_move()
-
-    def on_position_picked(self, callback):
-        self.emit(self.position_picked, callback)
-
-    def on_add_position(self):
-        dialog = PositionDialog()
-        dialog.positionPicked.connect(lambda dialog=dialog: self.on_position_picked(dialog.setPosition))  # TODO
-        if dialog.exec() == dialog.Accepted:
-            name = dialog.name()
-            x, y, z = dialog.position()
-            comment = dialog.comment()
-            self.positions_tree.append(TablePositionItem(name, x, y, z, comment))
-            self.positions_tree.fit()
-
-    def on_edit_position(self):
-        item = self.positions_tree.current
-        if item:
-            dialog = PositionDialog()
-            dialog.positionPicked.connect(lambda dialog=dialog: self.on_position_picked(dialog.setPosition))  # TODO
-            dialog.setName(item.name)
-            dialog.setPosition(*item.position)
-            dialog.setComment(item.comment)
-            if dialog.exec() == dialog.Accepted:
-                item.name = dialog.name()
-                item.position = dialog.position()
-                item.comment = dialog.comment()
-                self.positions_tree.fit()
-
-    def on_remove_position(self):
-        item = self.positions_tree.current
-        if item:
-            if ui.show_question(f"Do you want to remove position {item.name!r}?"):
-                self.positions_tree.remove(item)
-                if not len(self.positions_tree):
-                    self.edit_button.enabled = False
-                    self.remove_button.enabled = False
-                self.positions_tree.fit()
-
-    def on_move(self):
-        item = self.positions_tree.current
-        if item:
-            if ui.show_question(f"Do you want to move table to position {item.name!r}?"):
-                x, y ,z = item.position
-                self.emit(self.absolute_move, Position(x, y, z))
-
-
-class LCRWidget(QtWidgets.QWidget):
-
-    MaxPoints = 1000
-
-    def __init__(self, parent: Optional[QtWidgets.QWidget] = None) -> None:
-        super().__init__(parent)
-
-        self._chart = QtChart.QChart()
-        self._chart.legend().hide()
-        self._chart.layout().setContentsMargins(0, 0, 0, 0)
-        self._chart.setBackgroundRoundness(0)
-        self._chart.setBackgroundVisible(False)
-        self._chart.setMargins(QtCore.QMargins(0, 0, 0, 0))
-
-        self._xAxis = QtChart.QValueAxis()
-        self._xAxis.setTickCount(3)
-        self._xAxis.setMinorTickCount(4)
-        self._xAxis.setLabelFormat("%.3f mm")
-        self._chart.addAxis(self._xAxis, QtCore.Qt.AlignBottom)
-
-        self._yAxis = QtChart.QValueAxis()
-        self._yAxis.setTickCount(2)
-        self._yAxis.setMinorTickCount(3)
-        self._yAxis.setLabelFormat("%.2g Ohm")
-        self._chart.addAxis(self._yAxis, QtCore.Qt.AlignLeft)
-
-        self._line = QtChart.QLineSeries()
-        self._line.setColor(QtGui.QColor("magenta"))
-
-        self._chart.addSeries(self._line)
-        self._line.attachAxis(self._xAxis)
-        self._line.attachAxis(self._yAxis)
-
-        self._series = QtChart.QScatterSeries()
-        self._series.setName("R")
-        self._series.setMarkerSize(3)
-        self._series.setBorderColor(QtGui.QColor("red"))
-        self._series.setColor(QtGui.QColor("red"))
-
-        self._chart.addSeries(self._series)
-        self._series.attachAxis(self._xAxis)
-        self._series.attachAxis(self._yAxis)
-
-        self._marker = QtChart.QScatterSeries()
-        self._marker.setMarkerSize(9)
-        self._marker.setBorderColor(QtGui.QColor("red"))
-        self._marker.setColor(QtGui.QColor("red"))
-
-        self._chart.addSeries(self._marker)
-        self._marker.attachAxis(self._xAxis)
-        self._marker.attachAxis(self._yAxis)
-
-        self.setMinimumSize(160, 60)
-
-        self.chartView: QtChart.QChartView = QtChart.QChartView(self._chart)
-
-        layout = QtWidgets.QVBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.addWidget(self.chartView)
-
-    def yLimits(self) -> list:
-        limits: list = []
-        for point in self._series.pointsVector():
-            limits.append(point.y())
-        return limits
-
-    def clear(self) -> None:
-        self._series.clear()
-
-    def append(self, x: float, y: float) -> None:
-        if self._series.count() > type(self).MaxPoints:
-            self._series.remove(0)
-        self._series.append(QtCore.QPointF(x, y))
-        self.setMarker(x, y)
-
-    def setLimits(self, x: float) -> None:
-        self._xAxis.setRange(x - 0.050, x + 0.050)
-        limits = self.yLimits()
-        if limits:
-            self._yAxis.setRange(min(limits), max(limits))
-            self._yAxis.applyNiceNumbers()
-            self._yAxis.setTickCount(2)
-
-    def setLine(self, x: float) -> None:
-        self._line.clear()
-        self._line.append(x, self._yAxis.min())
-        self._line.append(x, self._yAxis.max())
-
-    def setMarker(self, x: float, y: float) -> None:
-        self._marker.clear()
-        self._marker.append(x, y)
-
-
-class TableControlDialog(QtWidgets.QDialog):
-
-    process = None
-
-    default_steps = [
-        {"step_size": 1.0, "step_color": "green"}, # microns!
-        {"step_size": 10.0, "step_color": "orange"},
-        {"step_size": 100.0, "step_color": "red"},
-    ]
-
-    maximum_z_step_size = 0.025 # mm
-    z_limit = 0.0
-
-    probecardLightToggled = QtCore.pyqtSignal(bool)
-    microscopeLightToggled = QtCore.pyqtSignal(bool)
-    boxLightToggled = QtCore.pyqtSignal(bool)
-
-    def __init__(self, process, lcr_process, parent: Optional[QtWidgets.QWidget] = None) -> None:
-        super().__init__(parent)
-
-        self.lcr_process = lcr_process
-        self.lcr_process.finished = self.on_lcr_finished
-        self.lcr_process.failed = self.on_lcr_failed
-        self.lcr_process.reading = self.on_lcr_reading
-
-        self.mount(process)
-
-        self.setWindowTitle("Table Control")
-
-        self.add_x_button = KeypadButton(
-            text="+X",
-            clicked=self.on_add_x
-        )
-        self.sub_x_button = KeypadButton(
-            text="-X",
-            clicked=self.on_sub_x
-        )
-        self.add_y_button = KeypadButton(
-            text="+Y",
-            clicked=self.on_add_y
-        )
-        self.sub_y_button = KeypadButton(
-            text="-Y",
-            clicked=self.on_sub_y
-        )
-        self.add_z_button = KeypadButton(
-            text="+Z",
-            clicked=self.on_add_z
-        )
-        self.sub_z_button = KeypadButton(
-            text="-Z",
-            clicked=self.on_sub_z
-        )
-        self.step_up_button = KeypadButton(
-            text="↑⇵",
-            tool_tip="Step up, move single step up then double step down and double step up (experimental).",
-            clicked=self.on_step_up
-        )
-
-        self.control_buttons = (
-            self.add_x_button,
-            self.sub_x_button,
-            self.add_y_button,
-            self.sub_y_button,
-            self.add_z_button,
-            self.sub_z_button,
-            self.step_up_button
-        )
-
-        self._lcr_prim_text = ui.Text(readonly=True)
-        self._lcr_sec_text = ui.Text(readonly=True)
-
-        self.lcrWidget: LCRWidget = LCRWidget(self)
-
-        self._lcr_groupbox = ui.GroupBox(
-            layout=ui.Column(
-                ui.Row(
-                    ui.Label("Cp"),
-                    self._lcr_prim_text
-                ),
-                ui.Row(
-                    ui.Label("Rp"),
-                    self._lcr_sec_text
-                )
-            )
-        )
-        self._lcr_groupbox.qt.setTitle("Contact Quality (LCR)")
-        self._lcr_groupbox.qt.setCheckable(True)
-        self._lcr_groupbox.qt.setChecked(False)
-        self._lcr_groupbox.qt.toggled.connect(self.on_lcr_toggled)
-        self._lcr_groupbox.qt.layout().addWidget(self.lcrWidget)
-
-        self.probecard_light_button = ToggleButton(
-            text="PC Light",
-            tool_tip="Toggle probe card light",
-            checkable=True,
-            checked=False,
-            enabled=False,
-            toggled=self.probecardLightToggled.emit
-        )
-        self.microscope_light_button = ToggleButton(
-            text="Mic Light",
-            tool_tip="Toggle microscope light",
-            checkable=True,
-            checked=False,
-            enabled=False,
-            toggled=self.microscopeLightToggled.emit
-        )
-        self.box_light_button = ToggleButton(
-            text="Box Light",
-            tool_tip="Toggle box light",
-            checkable=True,
-            checked=False,
-            enabled=False,
-            toggled=self.boxLightToggled.emit
-        )
-        # Create movement radio buttons
-        self.step_width_buttons = ui.Column()
-        for item in self.load_table_step_sizes():
-            step_size = item.get("step_size") * comet.ureg("um")
-            step_color = item.get("step_color")
-            step_size_label = format_metric(step_size.to("m").m, "m", decimals=1)
-            button = ui.RadioButton(
-                text=step_size_label,
-                tool_tip=f"Move in {step_size_label} steps.",
-                stylesheet=f"QRadioButton:enabled{{color:{step_color};}}",
-                checked=len(self.step_width_buttons) == 0,
-                toggled=self.on_step_toggled
-            )
-            button.movement_width = step_size.to("mm").m
-            button.movement_color = step_color
-            self.step_width_buttons.append(button)
-        self.control_layout = ui.Column(
-            ui.Row(
-                ui.Row(
-                    ui.Column(
-                        KeypadSpacer(),
-                        self.sub_y_button,
-                        KeypadSpacer()
-                    ),
-                    ui.Column(
-                        self.sub_x_button,
-                        KeypadSpacer(),
-                        self.add_x_button,
-                    ),
-                    ui.Column(
-                        KeypadSpacer(),
-                        self.add_y_button,
-                        KeypadSpacer(),
-                    ),
-                    KeypadSpacer(),
-                    ui.Column(
-                        ui.Row(
-                            self.add_z_button,
-                            self.step_up_button
-                        ),
-                        KeypadSpacer(),
-                        self.sub_z_button
-                    )
-                ),
-                ui.Spacer(),
-                stretch=(0, 1)
-            ),
-            ui.Spacer(),
-            stretch=(0, 1)
-        )
-        self.positions_widget = TablePositionsWidget(
-            enabled=False,
-            position_picked=self.on_position_picked,
-            absolute_move=self.on_absolute_move
-        )
-        self.contacts_widget = TableContactsWidget(
-            enabled=False,
-            position_picked=self.on_position_picked,
-            absolute_move=self.on_absolute_move
-        )
-        self._position_widget = PositionWidget()
-        self._calibration_widget = CalibrationWidget()
-        self.z_limit_label = PositionLabel()
-        self.x_hard_limit_label = PositionLabel()
-        self.y_hard_limit_label = PositionLabel()
-        self.z_hard_limit_label = PositionLabel()
-        self.laser_label = SwitchLabel()
-        self._calibrate_button = ui.Button(
-            text="Calibrate",
-            clicked=self.on_calibrate
-        )
-        self._update_interval_number = ui.Number(
-            value = settings.table_control_update_interval,
-            minimum=.5,
-            maximum=10.0,
-            decimals=2,
-            step=0.25,
-            suffix="s",
-            changed=self.on_update_interval_changed
-        )
-        self._interval_groupbox = ui.GroupBox(
-            title="Update Interval",
-            layout=ui.Row(
-                self._update_interval_number,
-                ui.Spacer(),
-                stretch=(0, 1)
-            )
-        )
-        self._dodge_height_number = ui.Number(
-            tool_tip="Doge height in microns.",
-            minimum=0,
-            maximum=10000,
-            decimals=0,
-            step=1,
-            suffix="um"
-        )
-        self._dodge_groupbox = ui.GroupBox(
-            title="X/Y Dodge",
-            tool_tip="Enables -/+ Z dodge for XY movements.",
-            checkable=True,
-            layout=ui.Row(
-                ui.Label("Height"),
-                self._dodge_height_number,
-                ui.Spacer(),
-                stretch=(0, 0, 1)
-            )
-        )
-        self._lcr_reset_on_move_checkbox = ui.CheckBox(
-            text="Reset graph on X/Y move"
-        )
-        self._contact_quality_groupbox = ui.GroupBox(
-            title="Contact Quality (LCR)",
-            layout=ui.Row(
-                self._lcr_reset_on_move_checkbox,
-                ui.Spacer(),
-                stretch=(0, 1)
-            )
-        )
-        self._step_up_delay_number = ui.Number(
-            minimum=0,
-            maximum=1000,
-            decimals=0,
-            step=25,
-            suffix="ms"
-        )
-        self._step_up_multiply_number = ui.Number(
-            minimum=1,
-            maximum=10,
-            decimals=0,
-            suffix="x"
-        )
-        self._step_up_groubox = ui.GroupBox(
-            title="Step Up (↑⇵)",
-            layout=ui.Row(
-                ui.Column(
-                    ui.Label("Delay"),
-                    ui.Label("Multiplicator (⇵)")
-                ),
-                ui.Column(
-                    self._step_up_delay_number,
-                    self._step_up_multiply_number
-                ),
-                ui.Spacer()
-            )
-        )
-        self._lcr_update_interval_number = ui.Number(
-            minimum=0,
-            maximum=1000,
-            decimals=0,
-            step=25,
-            suffix="ms"
-        )
-        self._lcr_matrix_channels_text = ui.Text(
-        )
-        self._lcr_options_groupbox = ui.GroupBox(
-            title="Contact Quality (LCR)",
-            layout=ui.Row(
-                ui.Column(
-                    ui.Label("Reading Interval"),
-                    ui.Label("Matrix Channels"),
-                ),
-                ui.Column(
-                    ui.Row(
-                        self._lcr_update_interval_number,
-                        ui.Spacer()
-                    ),
-                    self._lcr_matrix_channels_text
-                )
-            )
-        )
-
-        self.progressBar: QtWidgets.QProgressBar = QtWidgets.QProgressBar(self)
-        self.progressBar.setVisible(False)
-
-        self.messageLabel: QtWidgets.QLabel = QtWidgets.QLabel(self)
-
-        self.stopButton: QtWidgets.QPushButton = QtWidgets.QPushButton(self)
-        self.stopButton.setText("&Stop")
-        self.stopButton.setDefault(False)
-        self.stopButton.setAutoDefault(False)
-        self.stopButton.setEnabled(False)
-        self.stopButton.clicked.connect(self.stop)
-
-        self.closeButton: QtWidgets.QPushButton = QtWidgets.QPushButton(self)
-        self.closeButton.setText("&Close")
-        self.closeButton.setDefault(False)
-        self.closeButton.setAutoDefault(False)
-        self.closeButton.clicked.connect(self.close)
-
-        self.control_panel = ui.Row(
-            ui.Column(
-                ui.Row(
-                    ui.GroupBox(
-                        title="Control",
-                        layout=ui.Column(
-                            ui.Spacer(horizontal=False),
-                            self.control_layout,
-                            ui.Spacer(horizontal=False)
-                        )
-                    ),
-                    ui.GroupBox(
-                        title="Step Width",
-                        layout=self.step_width_buttons
-                    ),
-                    self._lcr_groupbox,
-                    ui.Column(
-                        ui.GroupBox(
-                            title="Lights",
-                            layout=ui.Column(
-                                self.probecard_light_button,
-                                self.microscope_light_button,
-                                self.box_light_button,
-                                ui.Spacer()
-                            )
-                        )
-                    ),
-                    stretch=(0, 0, 1, 0)
-                ),
-                ui.Tabs(
-                    ui.Tab(
-                        title="Move",
-                        layout=self.positions_widget
-                    ),
-                    ui.Tab(
-                        title="Contacts",
-                        layout=self.contacts_widget
-                    ),
-                    ui.Tab(
-                        title="Calibrate",
-                        layout=ui.Column(
-                            ui.GroupBox(
-                                title="Table Calibration",
-                                layout=ui.Column(
-                                    ui.Row(
-                                        self._calibrate_button,
-                                        ui.Label("Calibrate table by moving into cal/rm switches\nof every axis in" \
-                                                    " a safe manner to protect the probe card."),
-                                        stretch=(0, 1)
-                                    )
-                                )
-                            ),
-                            ui.Spacer(),
-                            stretch=(0, 1)
-                        )
-                    ),
-                    ui.Tab(
-                        title="Options",
-                        layout=ui.Column(
-                            ui.Row(
-                                self._interval_groupbox,
-                                self._dodge_groupbox,
-                                self._contact_quality_groupbox,
-                                stretch=(0, 0, 1)
-                            ),
-                            self._step_up_groubox,
-                            self._lcr_options_groupbox,
-                            ui.Spacer(),
-                            stretch=(0, 0, 0, 1)
-                        )
-                    )
-                ),
-                stretch=(0, 1)
-            ),
-            ui.Column(
-                self._position_widget,
-                self._calibration_widget,
-                ui.GroupBox(
-                    title="Soft Limits",
-                    layout=ui.Row(
-                        ui.Column(
-                            ui.Label("Z")
-                        ),
-                        ui.Column(
-                            self.z_limit_label
-                        )
-                    )
-                ),
-                ui.GroupBox(
-                    title="Hard Limits",
-                    layout=ui.Row(
-                        ui.Column(
-                            ui.Label("X"),
-                            ui.Label("Y"),
-                            ui.Label("Z")
-                        ),
-                        ui.Column(
-                            self.x_hard_limit_label,
-                            self.y_hard_limit_label,
-                            self.z_hard_limit_label
-                        )
-                    )
-                ),
-                ui.GroupBox(
-                    title="Safety",
-                    layout=ui.Row(
-                        ui.Label(
-                            text="Laser Sensor"
-                        ),
-                        self.laser_label
-                    )
-                ),
-                ui.Spacer()
-            ),
-            stretch=(1, 0)
-        )
-
-        self.buttomLayout = QtWidgets.QHBoxLayout()
-        self.buttomLayout.addWidget(self.progressBar)
-        self.buttomLayout.addWidget(self.messageLabel)
-        self.buttomLayout.addStretch()
-        self.buttomLayout.addWidget(self.stopButton)
-        self.buttomLayout.addWidget(self.closeButton)
-
-        layout = QtWidgets.QVBoxLayout(self)
-        layout.addWidget(self.control_panel.qt, 1)
-        layout.addLayout(self.buttomLayout, 0)
-
-        self.resetPosition()
-        self.resetCaldone()
-        self.updateLimits()
-        self.resetSafety()
-        self.update_control_buttons()
-
-    @property
-    def step_width(self):
-        for button in self.step_width_buttons:
-            if button.checked:
-                return abs(button.movement_width)
-        return 0
-
-    @property
-    def step_color(self):
-        for button in self.step_width_buttons:
-            if button.checked:
-                return button.movement_color
-        return "black"
-
-    @property
-    def update_interval(self):
-        return self._update_interval_number.value
-
-    @update_interval.setter
-    def update_interval(self, value):
-        self._update_interval_number.value = value
-
-    @property
-    def dodge_enabled(self):
-        return self._dodge_groupbox.checked
-
-    @dodge_enabled.setter
-    def dodge_enabled(self, value):
-        self._dodge_groupbox.checked = value
-
-    @property
-    def dodge_height(self):
-        return (self._dodge_height_number.value * comet.ureg("um")).to("mm").m
-
-    @dodge_height.setter
-    def dodge_height(self, value):
-        self._dodge_height_number.value = (value * comet.ureg("mm")).to("um").m
-
-    @property
-    def lcr_reset_on_move(self):
-        return self._lcr_reset_on_move_checkbox.checked
-
-    @lcr_reset_on_move.setter
-    def lcr_reset_on_move(self, value):
-        self._lcr_reset_on_move_checkbox.checked = value
-
-    @property
-    def step_up_delay(self):
-        """Return step up delay in seconds."""
-        return (self._step_up_delay_number.value * comet.ureg("ms")).to("s").m
-
-    @step_up_delay.setter
-    def step_up_delay(self, value):
-        self._step_up_delay_number.value = (value * comet.ureg("s")).to("ms").m
-
-    @property
-    def step_up_multiply(self):
-        """Return step up delay in seconds."""
-        return int(self._step_up_multiply_number.value)
-
-    @step_up_multiply.setter
-    def step_up_multiply(self, value):
-        self._step_up_multiply_number.value = value
-
-    @property
-    def lcr_update_interval(self):
-        """LCR update interval in seconds."""
-        return (self._lcr_update_interval_number.value * comet.ureg("ms")).to("s").m
-
-    @lcr_update_interval.setter
-    def lcr_update_interval(self, value):
-        self._lcr_update_interval_number.value = (value * comet.ureg("s")).to("ms").m
-
-    @property
-    def lcr_matrix_channels(self):
-        """Matrix channels used for LCR readings."""
-        tokens = []
-        for token in self._lcr_matrix_channels_text.value.split(","):
-            token = token.strip()
-            if token:
-                tokens.append(token)
-        return tokens
-
-    @lcr_matrix_channels.setter
-    def lcr_matrix_channels(self, value):
-        self._lcr_matrix_channels_text.value = ", ".join([token for token in value])
-
-    def load_table_step_sizes(self):
-        return settings.settings.get("table_step_sizes") or self.default_steps
-
-    def resetPosition(self) -> None:
-        self.setPosition(Position())
-
-    def setPosition(self, position: Position) -> None:
-        self.current_position = position
-        self._position_widget.update_position(position)
-        self.updateLimits()
-        self.update_control_buttons()
-        if math.isfinite(position.z):
-            self.lcrWidget.setLimits(position.z)
-            self.lcrWidget.setLine(position.z)
-
-    def resetCaldone(self) -> None:
-        self._calibration_widget.reset_calibration()
-
-    def setCaldone(self, position: Position) -> None:
-        self.current_caldone = position
-        self.positions_widget.enabled = caldone_valid(position)
-        self.contacts_widget.enabled = caldone_valid(position)
-        self.control_layout.enabled = caldone_valid(position)
-        self._calibration_widget.update_calibration(position)
-
-    def updateLimits(self) -> None:
-        x, y, z = self.current_position
-        self.z_limit_label.stylesheet = ""
-        if not math.isnan(z):
-            if z >= self.z_limit:
-                self.z_limit_label.stylesheet = "QLabel:enabled{color:red;}"
-
-    def resetSafety(self) -> None:
-        self.laser_label.setValue(None)
-
-    def updateSafety(self, value) -> None:
-        self.laser_label.setValue(value)
-
-    def update_control_buttons(self):
-        x, y, z = self.current_position
-        self.update_x_buttons(x)
-        self.update_y_buttons(y)
-        self.update_z_buttons(z)
-        for button in self.control_buttons:
-            button.stylesheet = f"QPushButton:enabled{{color:{self.step_color or 'black'}}}"
-
-    def update_x_buttons(self, x):
-        x_enabled = True
-        if not math.isnan(x):
-            if (x - self.step_width) < 0:
-                x_enabled = False
-        self.sub_x_button.enabled = x_enabled
-
-    def update_y_buttons(self, y):
-        y_enabled = True
-        if not math.isnan(y):
-            if (y - self.step_width) < 0:
-                y_enabled = False
-        self.sub_y_button.enabled = y_enabled
-
-    def update_z_buttons(self, z):
-        # Disable move up button for large step sizes
-        z_enabled = False
-        if not math.isnan(z):
-            if (z + self.step_width) <= self.z_limit:
-                z_enabled = True
-            else:
-                z_enabled = self.step_width <= self.maximum_z_step_size
-        self.add_z_button.enabled = z_enabled
-        step_up_limit = comet.ureg("10.0 um").to("mm").m
-        self.step_up_button.enabled = z_enabled and (self.step_width <= step_up_limit) # TODO
-
-    def relative_move_xy(self, x, y):
-        # Dodge on X/Y movements.
-        if self.dodge_enabled:
-            dodge_height = self.dodge_height
-            current_position = self.current_position
-            if current_position.z < dodge_height:
-                dodge_height = max(0, current_position.z)
-            vector = [(0, 0, -dodge_height), (x, y, 0), (0, 0, +dodge_height)]
-        else:
-            vector = [(x, y, 0)]
-        # Clear contact quality graph on X/Y movements.
-        if self.lcr_reset_on_move:
-            self.lcrWidget.clear()
-        self.process.relative_move_vector(vector)
-        # Clear contact quality graph on X/Y movements.
-        if self.lcr_reset_on_move:
-            self.lcrWidget.clear()
-
-    def on_add_x(self):
-        self.setLocked(True)
-        self.relative_move_xy(+self.step_width, 0)
-
-    def on_sub_x(self):
-        self.setLocked(True)
-        self.relative_move_xy(-self.step_width, 0)
-
-    def on_add_y(self):
-        self.setLocked(True)
-        self.relative_move_xy(0, +self.step_width)
-
-    def on_sub_y(self):
-        self.setLocked(True)
-        self.relative_move_xy(0, -self.step_width)
-
-    def on_add_z(self):
-        self.setLocked(True)
-        self.process.relative_move(0, 0, +self.step_width)
-
-    def on_sub_z(self):
-        self.setLocked(True)
-        self.process.relative_move(0, 0, -self.step_width)
-
-    def on_step_up(self):
-        self.setLocked(True)
-        step_width = self.step_width
-        multiply = self.step_up_multiply
-        vector = (
-            [0, 0, +step_width],
-            [0, 0, -step_width * multiply],
-            [0, 0, +step_width * multiply],
-        )
-        self.process.relative_move_vector(vector, delay=self.step_up_delay)
-
-    def on_step_toggled(self, state):
-        logger.info("set table step width to %.3f mm", self.step_width)
-        self.update_control_buttons()
-
-    def update_probecard_light(self, state):
-        self.probecard_light_button.checked = state
-
-    def update_microscope_light(self, state):
-        self.microscope_light_button.checked = state
-
-    def update_box_light(self, state):
-        self.box_light_button.checked = state
-
-    def update_lights_enabled(self, state):
-        self.probecard_light_button.enabled = state
-        self.microscope_light_button.enabled = state
-        self.box_light_button.enabled = state
-
-    def on_move_finished(self):
-        self.progressBar.setVisible(False)
-        self.stopButton.setEnabled(False)
-        self.setLocked(False)
-
-    def on_calibration_finished(self):
-        self.progressBar.setVisible(False)
-        self.stopButton.setEnabled(False)
-        self.setLocked(False)
-
-    def setMessage(self, message: str) -> None:
-        self.messageLabel.setText(message)
-        logger.info(message)
-
-    def setProgress(self, value: int, maximum: int) -> None:
-        self.progressBar.setValue(value)
-        self.progressBar.setMaximum(maximum)
-        self.progressBar.setVisible(True)
-
-    def on_update_interval_changed(self, value):
-        self.process.update_interval = value
-
-    def on_position_picked(self, callback):
-        x, y, z = self.current_position
-        callback(x, y, z)
-
-    def on_absolute_move(self, position):
-        # Update to safe Z position
-        position = Position(position.x, position.y, safe_z_position(position.z))
-        self.setLocked(True)
-        # Clear contact quality graph on X/Y movements.
-        self.lcrWidget.clear()
-        self.stopButton.setEnabled(True)
-        self.process.safe_absolute_move(position.x, position.y, position.z)
-
-    def on_calibrate(self):
-        self.setLocked(True)
-        self.stopButton.setEnabled(True)
-        self.process.calibrate_table()
-
-    def stop(self) -> None:
-        self.stopButton.setEnabled(False)
-        self.process.stop_current_action()
-
-    def closeEvent(self, event: QtCore.QEvent) -> None:
-        self.process.wait()
-        event.accept()
-
-    def loadSamples(self, sample_items):
-        self.contacts_widget.loadSamples(sample_items)
-
-    def update_samples(self):
-        self.contacts_widget.updateSamples()
-
-    def readSettings(self) -> None:
-        settings_ = QtCore.QSettings()
-        settings_.beginGroup("tablecontrol")
-        geometry = settings_.value("geometry", QtCore.QByteArray(), QtCore.QByteArray)
-        if not geometry.isEmpty():
-            self.restoreGeometry(geometry)
-        settings_.endGroup()
-
-        self.positions_widget.readSettings()
-        self.z_limit = settings.table_z_limit
-        self.z_limit_label.value = self.z_limit
-        x, y, z = settings.table_probecard_maximum_limits
-        self.x_hard_limit_label.value = x
-        self.y_hard_limit_label.value = y
-        self.z_hard_limit_label.value = z
-        self.step_up_delay = settings.settings.get("tablecontrol_step_up_delay", DEFAULT_STEP_UP_DELAY)
-        self.step_up_multiply = settings.settings.get("tablecontrol_step_up_multiply", DEFAULT_STEP_UP_MULTIPLY)
-        self.lcr_update_interval = settings.settings.get("tablecontrol_lcr_update_delay", DEFAULT_LCR_UPDATE_INTERVAL)
-        matrix_channels = settings.settings.get("tablecontrol_lcr_matrix_channels") or DEFAULT_MATRIX_CHANNELS
-        self.lcr_matrix_channels = matrix_channels
-        self.lcr_process.update_interval = self.lcr_update_interval
-        self.lcr_process.matrix_channels = self.lcr_matrix_channels
-        self.update_interval = settings.table_control_update_interval
-        self.dodge_enabled = settings.table_control_dodge_enabled
-        self.dodge_height = settings.table_control_dodge_height
-        self.lcr_reset_on_move = settings.settings.get("tablecontrol_lcr_reset_on_move", True)
-
-    def writeSettings(self) -> None:
-        settings_ = QtCore.QSettings()
-        settings_.beginGroup("tablecontrol")
-        settings_.setValue("geometry", self.saveGeometry())
-        settings_.endGroup()
-
-        settings.settings["tablecontrol_step_up_multiply"] = self.step_up_multiply
-        settings.settings["tablecontrol_lcr_update_delay"] = self.lcr_update_interval
-        settings.settings["tablecontrol_lcr_matrix_channels"] = self.lcr_matrix_channels
-        self.positions_widget.writeSettings()
-        settings.table_control_update_interval = self.update_interval
-        settings.table_control_dodge_enabled = self.dodge_enabled
-        settings.table_control_dodge_height = self.dodge_height
-        settings.settings["tablecontrol_lcr_reset_on_move"] = self.lcr_reset_on_move
-
-    def setLocked(self, locked: bool) -> None:
-        self.control_layout.qt.setEnabled(not locked)
-        self.positions_widget.setLocked(locked)
-        self.contacts_widget.setLocked(locked)
-        self.closeButton.setEnabled(not locked)
-        self.progressBar.setVisible(locked)
-        if locked:
-            self.progressBar.setRange(0, 0)
-            self.progressBar.setValue(0)
-
-    def mount(self, process):
-        """Mount table process."""
-        self.unmount()
-        self.process = process
-        self.process.message_changed = self.setMessage
-        self.process.progress_changed = self.setProgress
-        self.process.position_changed = self.setPosition
-        self.process.caldone_changed = self.setCaldone
-        self.process.relative_move_finished = self.on_move_finished
-        self.process.absolute_move_finished = self.on_move_finished
-        self.process.calibration_finished = self.on_calibration_finished
-        self.process.stopped = self.on_calibration_finished
-
-    def unmount(self):
-        """Unmount table process."""
-        if self.process:
-            self.process.message_changed = None
-            self.process.progress_changed = None
-            self.process.position_changed = None
-            self.process.caldone_changed = None
-            self.process.relative_move_finished = None
-            self.process.absolute_move_finished = None
-            self.process.stopped = None
-            self.process = None
-
-    @handle_exception
-    def on_lcr_toggled(self, state):
-        self._lcr_prim_text.enabled = state
-        self._lcr_sec_text.enabled = state
-        self.lcrWidget.setEnabled(state)
-        if state:
-            self.lcrWidget.clear()
-            self.lcr_process.update_interval = self.lcr_update_interval
-            self.lcr_process.matrix_channels = self.lcr_matrix_channels
-            self.lcr_process.start()
-        else:
-            self.lcr_process.stop()
-
-    def on_lcr_finished(self):
-        ...
-
-    def on_lcr_failed(self, *args):
-        self._lcr_prim_text.value = "ERROR"
-        self._lcr_sec_text.value = "ERROR"
-
-    def on_lcr_reading(self, prim, sec):
-        self._lcr_prim_text.value = format_metric(prim, unit="F")
-        self._lcr_sec_text.value = format_metric(sec, unit="Ohm")
-        _, _, z = self.current_position
-        if math.isfinite(z) and math.isfinite(sec):
-            # Append only absolute Rp readings
-            self.lcrWidget.append(z, abs(sec))
-            self.lcrWidget.setLine(z)
-
-
-class SwitchLabel(ui.Label):
-
-    def __init__(self, value=None):
-        super().__init__()
-        self.setValue(value)
-
-    def setValue(self, value) -> None:
-        if value is None:
-            self.qt.setText(format(float("nan")))
-            self.qt.setStyleSheet("")
-        else:
-            self.qt.setText(format_switch(value))
-            self.qt.setStyleSheet("QLabel:enabled{color:green}" if value else "QLabel:enabled{color:red}")
-
-
-class KeypadSpacer(ui.Spacer):
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.qt.setFixedSize(30, 30)
-
-
-class KeypadButton(ui.Button):
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.qt.setFixedSize(32, 32)
